refactor: remove unused AI_TOOL_INDEX imports and improve type definitions in generateCustomFetchTool
