+++ app/api/v1/chat/[chatId]/route.ts
@@ -8,7 +8,6 @@ import { z } from 'zod'
 import { openai } from '@ai-sdk/openai'
 import { ObjectId } from 'bson'
 import { saveMessages } from '@/data/chat.server'
-import { AI_TOOL_INDEX } from '@/ai_tools'
 import { Prisma } from '@prisma/client'
 import { ModelId, modelPrices } from '@/lib/constants/models'
 import { db } from '@/lib/db'
