+++ app/api/v1/chat/[chatId]/stream/route.ts
@@ -15,7 +15,6 @@ import { z } from 'zod'
 import { openai } from '@ai-sdk/openai'
 import { ObjectId } from 'bson'
 import { saveMessages } from '@/data/chat.server'
-import { AI_TOOL_INDEX } from '@/ai_tools'
 import { Prisma } from '@prisma/client'
 import { ModelId, modelPrices } from '@/lib/constants/models'
 import { generateTools } from '@/lib/ai'
